Move vvh plotting to vvh analysis dir
