+++ analysis/vbs_vvh/check_vbs_hists.py
@@ -18,7 +18,6 @@ import ewkcoffea.modules.yield_tools as yt
 import ewkcoffea.modules.sample_groupings as sg
 import ewkcoffea.modules.plotting_tools as plt_tools
 
-import get_wwz_yields as gwy
 
 HTML_PC = "/home/users/kmohrman/ref_scripts/html_stuff/index.php"
 
@@ -165,6 +164,75 @@ def append_years(sample_dict_base,year_lst):
                 out_dict[proc_group].append(f"{year_str}_{proc_base_name}")
     return out_dict
 
+##################################################################
+# Make the figures for the vvh sudy
+def make_vvh_fig(histo_mc,histo_mc_sig,histo_mc_bkg,title="test",axisrangex=None):
+
+    # Create the figure
+    fig, (ax1, ax2) = plt.subplots(
+        nrows=2,
+        ncols=1,
+        figsize=(7,7),
+        gridspec_kw={"height_ratios": (3, 1)},
+        sharex=True
+    )
+    fig.subplots_adjust(hspace=.07)
+
+    # Plot the stack plot
+    histo_mc.plot1d(
+        stack=True,
+        histtype="fill",
+        #color=CLR_LST,
+        ax=ax1,
+    )
+
+    # Get the errs on MC and plot them by hand on the stack plot
+    histo_mc_sum = histo_mc[{"process_grp":sum}]
+    mc_arr = histo_mc_sum.values()
+    mc_err_arr = np.sqrt(histo_mc_sum.variances())
+    err_p = np.append(mc_arr + mc_err_arr, 0)
+    err_m = np.append(mc_arr - mc_err_arr, 0)
+    bin_edges_arr = histo_mc_sum.axes[0].edges
+    bin_centers_arr = histo_mc_sum.axes[0].centers
+    ax1.fill_between(bin_edges_arr,err_m,err_p, step='post', facecolor='none', edgecolor='gray', alpha=0.5, linewidth=0.0, label='MC stat', hatch='/////')
+
+
+    # Get normalized hists of sig and bkg
+    yld_sig = sum(sum(histo_mc_sig.values(flow=True)))
+    yld_bkg = sum(sum(histo_mc_bkg.values(flow=True)))
+    histo_mc_sig_scale_to_bkg = plt_tools.scale(copy.deepcopy(histo_mc_sig), "process_grp", {"Signal":yld_bkg/yld_sig})
+    histo_mc_sig_norm         = plt_tools.scale(copy.deepcopy(histo_mc_sig), "process_grp", {"Signal":1.0/yld_sig})
+    histo_mc_bkg_norm         = plt_tools.scale(copy.deepcopy(histo_mc_bkg), "process_grp", {"Background":1.0/yld_bkg})
+
+    # Plot the normalized shapes
+    histo_mc_sig_scale_to_bkg.plot1d(color=["red"], ax=ax1,)
+    histo_mc_sig_norm.plot1d(color="red",  ax=ax2,)
+    histo_mc_bkg_norm.plot1d(color="gray", ax=ax2,)
+
+    # Draw legend, scale the axis, set labels, etc
+    extr = ax1.legend(loc="upper left", bbox_to_anchor=(1, 1), fontsize="12", frameon=False)
+    extr = ax2.legend(loc="upper left", bbox_to_anchor=(1, 1), fontsize="12", frameon=False)
+    plt.text(0.15,0.85, f"[Signal overlay scaled by {np.round(yld_bkg/yld_sig,2)}]", fontsize = 12, transform=fig.transFigure)
+
+    extt = ax1.set_title(title)
+    extb = ax1.set_xlabel(None)
+    extl = ax2.set_ylabel('Shapes')
+    ax1.tick_params(axis='y', labelsize=16)
+    ax2.tick_params(axis='x', labelsize=16)
+
+    shapes_ymax = max( max(sum(histo_mc_sig_norm.values(flow=True))) , max(sum(histo_mc_bkg_norm.values(flow=True))) )
+    ax2.set_ylim(0.0,1.8*shapes_ymax)
+    ax1.autoscale(axis='y')
+    #ax1.set_yscale('log')
+
+    if axisrangex is not None:
+        ax1.set_xlim(axisrangex[0],axisrangex[1])
+        ax2.set_xlim(axisrangex[0],axisrangex[1])
+
+
+    return (fig,(extt,extr,extb,extl))
+    ##################################################################
+
 
 def main():
 
@@ -266,7 +334,7 @@ def main():
 
                 # Make the figure
                 title = f"{cat}__{var}"
-                fig,ext_tup = gwy.make_vvh_fig(
+                fig,ext_tup = make_vvh_fig(
                     histo_mc = histo,
                     histo_mc_sig = histo_sig,
                     histo_mc_bkg = histo_bkg,
